requirements.txt
telethon==1.34.0
pandas==2.2.1
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
joblib==1.3.2 